# pkg/presentation/http/intern/downloads-openapi.yaml
openapi: 3.0.3
info:
  title: Download solution archives for MetalK8s
  description: "Download solution archives for MetalK8s"
  version: 1.0.0
servers:
  - url: https://IP:5002/api/v1/
paths:
  /downloads/{solutionArchive}/{version}:
    get:
      summary: Download the solution archive.
      operationId: downloadSolutionArchive
      parameters:
        - name: solutionArchive
          in: path
          description: Name of the solution archive
          required: true
          schema:
            type: string
        - name: version
          in: path
          description: Version of the solution archive
          required: true
          schema:
            type: string
        - name: Range
          in: header
          description: >-
            Range of the chunk to serve. Must respect the format `bytes=<start-index>-<end-index>`
          required: true
          example: bytes=0-2
          schema:
            type: string
      responses:
        "206":
          description: successful partial content
          headers:
            Content-Range:
              description: The range of the chunk served
              required: true
              schema:
                type: string
                format: bytes <start-index>-<end-index>/<total-size-in-bytes>
                example: bytes 0-2/4
            # The 'Content-Digest' header is sent as an HTTP trailer.
            # As oapi-codegen does not handle this feature, we need to manually reimplement
            # the VisitDownloadSolutionArchiveResponse method
            # (See pkg/presentation/http/handler/download_solution_archive.go).
            Content-Digest:
              description: The digest of the chunk, Base64 encoded
              required: true
              schema:
                type: string
                format: base64
                example: "sha-256=:d435Qo+nKZ+gLcUHn7GQtQ72hiBVAgqoLsZnZPiTGPk=:"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/ServerError"
    head:
      summary: Describe the solution archive.
      operationId: describeSolutionArchive
      parameters:
        - name: solutionArchive
          in: path
          description: Name of the solution archive
          required: true
          schema:
            type: string
        - name: version
          in: path
          description: Version of the solution archive
          required: true
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          headers:
            Content-Length:
              description: Size of the requested file in bytes
              required: true
              schema:
                type: integer
                format: int64
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
components:
  responses:
    BadRequest:
      description: |-
        The request was malformed or could not be processed. 

        Examples of `Bad Request` problem detail responses:

         - [Missing body property](https://problems-registry.smartbear.com/missing-body-property/)
         - [Invalid body property format](https://problems-registry.smartbear.com/invalid-body-property-format/)
         - [Invalid body property value](https://problems-registry.smartbear.com/invalid-body-property-value/)
         - [Missing request header](https://problems-registry.smartbear.com/missing-request-header/)
         - [Invalid request header format](https://problems-registry.smartbear.com/invalid-request-header-format/)
         - [Missing request parameter](https://problems-registry.smartbear.com/missing-request-parameter/)         
         - [Invalid request parameter format](https://problems-registry.smartbear.com/invalid-request-parameter-format/)
         - [Invalid request parameter value](https://problems-registry.smartbear.com/invalid-request-parameter-value/)
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            missingBodyProperty:
              $ref: "#/components/examples/missing-body-property"
            missingBodyPropertyWithErrors:
              $ref: "#/components/examples/missing-body-property-with-errors"
            missingRequestParameter:
              $ref: "#/components/examples/missing-request-parameter"
            missingRequestParameterWithErrors:
              $ref: "#/components/examples/missing-request-parameter-with-errors"
            missingRequestHeader:
              $ref: "#/components/examples/missing-request-header"
            missingRequestHeaderWithErrors:
              $ref: "#/components/examples/missing-request-header-with-errors"
            invalidBodyPropertyFormat:
              $ref: "#/components/examples/invalid-body-property-format"
            invalidBodyPropertyFormatWithErrors:
              $ref: "#/components/examples/invalid-body-property-format-with-errors"
            invalidRequestHeaderFormat:
              $ref: "#/components/examples/invalid-request-header-format"
            invalidRequestHeaderFormatWithErrors:
              $ref: "#/components/examples/invalid-request-header-format-with-errors"
            invalidBodyPropertyValue:
              $ref: "#/components/examples/invalid-body-property-value"
            invalidBodyPropertyValueWithErrors:
              $ref: "#/components/examples/invalid-body-property-value-with-errors"
    Unauthorized:
      description: Access token is not set or invalid.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            unauthorized:
              $ref: "#/components/examples/unauthorized"
    Forbidden:
      description: The requestor is not authorized to perform this operation on the resource.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            forbidden:
              $ref: "#/components/examples/forbidden"
    NotFound:
      description: The requested resource was not found.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            notFound:
              $ref: "#/components/examples/not-found"
    UnprocessableEntity:
      description: The request was malformed or could not be processed. 
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            unprocessableEntity:
              $ref: "#/components/examples/unprocessable-entity"
    ServerError:
      description: The server encountered an unexpected error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            serverError:
              $ref: "#/components/examples/server-error"
  schemas:
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          description: A URI reference that identifies the problem type.
          format: uri
          maxLength: 1024
        status:
          type: integer
          description: The HTTP status code generated by the origin server for this occurrence of the problem.
          format: int32
          minimum: 100
          maximum: 599
        title:
          type: string
          description: >-
            A short, human-readable summary of the problem type. It should not change from occurrence to occurrence of the problem, except for purposes of localization.
          maxLength: 1024
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
          maxLength: 4096
        instance:
          type: string
          description: >-
            A URI reference that identifies the specific occurrence of the problem. It may or may not yield further information if dereferenced.
          maxLength: 1024
        code:
          type: string
          description: >-
            An API specific error code aiding the provider team understand the error based on their own potential taxonomy or registry.
          maxLength: 50
        additionalInformation:
          type: object
          description: A map containing flexible metadata or contextual details about the error.
          additionalProperties: true # Allows keys of string to map to any type of value
          example:
            invalidFields: ["email", "zipCode"]
            timestamp: "2026-05-21T08:14:00Z"
            internalId: 98765
      required:
        - title
  examples:
    missing-body-property:
      description: Representation of a 400 Bad Request response where a rquired `body` property was omitted by the consumer.
      value:
        type: https://problems-registry.smartbear.com/missing-body-property
        status: 400
        title: Missing body property
        detail: The request is missing an expected body property.
        code: 400-09
    missing-body-property-with-errors:
      description: Representation of a 400 Bad Request response where a rquired `body` property was omitted by the consumer.
      value:
        type: https://problems-registry.smartbear.com/missing-body-property
        status: 400
        title: Missing body property
        detail: The request is missing an expected body property.
        code: 400-09
    missing-request-parameter:
      description: Representation of a 400 Bad Request response where a `query` or `path` parameter was omitted by the consumer.
      value:
        type: https://problems-registry.smartbear.com/missing-request-parameter
        status: 400
        title: Missing request parameter
        detail: The request is missing an expected query or path parameter.
        code: 400-03
    missing-request-parameter-with-errors:
      description: Representation of a 400 Bad Request response where a `query` or `path` parameter was omitted by the consumer.
      value:
        type: https://problems-registry.smartbear.com/missing-request-parameter
        status: 400
        title: Missing request parameter
        detail: The request is missing an expected query or path parameter.
        code: 400-03
    missing-request-header:
      description: Representation of a 400 Bad Request response where a `header` parameter was omitted by the consumer.
      value:
        type: https://problems-registry.smartbear.com/missing-request-header
        status: 400
        title: Missing request header
        detail: The request is missing an expected HTTP request header.
        code: 400-02
    missing-request-header-with-errors:
      description: Representation of a 400 Bad Request response where a `header` parameter was omitted by the consumer.
      value:
        type: https://problems-registry.smartbear.com/missing-request-header
        status: 400
        title: Missing request header
        detail: The request is missing an expected HTTP request header.
        code: 400-02
    invalid-body-property-format:
      description: Representation of a 400 Bad Request response where a body property does not leverage the correct format
      value:
        type: https://problems-registry.smartbear.com/invalid-body-property-format
        status: 400
        title: Invalid Body Property Format
        detail: The request body contains a malformed property.
        code: 400-04
    invalid-body-property-format-with-errors:
      description: Representation of a 400 Bad Request response where a body property does not leverage the correct format
      value:
        type: https://problems-registry.smartbear.com/invalid-body-property-format
        status: 400
        title: Invalid Body Property Format
        detail: The request body contains a malformed property.
        code: 400-04
    invalid-request-parameter-format:
      description: Representation of a 400 Bad Request response where a query or path parameter is malformed.
      value:
        type: https://problems-registry.smartbear.com/invalid-request-parameter-format
        status: 400
        title: Invalid Request Parameter Format
        detail: The request contains a malformed query parameter.
        code: 400-05
    invalid-request-parameter-format-with-errors:
      description: Representation of a 400 Bad Request response where a query or path parameter is malformed.
      value:
        type: https://problems-registry.smartbear.com/invalid-request-parameter-format
        status: 400
        title: Invalid Request Parameter Format
        detail: The request contains a malformed query parameter.
        code: 400-05
    invalid-request-header-format:
      description: Representation of a 400 Bad Request response where request header provided was malformed
      value:
        type: https://problems-registry.smartbear.com/invalid-request-header-format
        status: 400
        title: Invalid Request Header Format
        detail: The request contains a malformed request header parameter.
        code: 400-06
    invalid-request-header-format-with-errors:
      description: Representation of a 400 Bad Request response where a query or path parameter is malformed.
      value:
        type: https://problems-registry.smartbear.com/invalid-request-header-format
        status: 400
        title: Invalid Request Header Format
        detail: The request contains a malformed request header parameter.
        code: 400-06
    invalid-body-property-value:
      description: Representation of a 400 Bad Request response where the request body contains a invalid property value
      value:
        type: https://problems-registry.smartbear.com/invalid-body-property-value
        status: 400
        title: Invalid Body Property Value
        detail: The request body contains an invalid body property value.
        code: 400-07
    invalid-body-property-value-with-errors:
      description: Representation of a 400 Bad Request response where the request body contains a invalid property value
      value:
        type: https://problems-registry.smartbear.com/invalid-body-property-value
        status: 400
        title: Invalid Body Property Value
        detail: The request body contains an invalid body property value.
        code: 400-07
    invalid-request-parameter-value:
      description: Representation of a 400 Bad Request response where the request contains a invalid query or path parameter value.
      value:
        type: https://problems-registry.smartbear.com/invalid-request-parameter-value
        status: 400
        title: Invalid Request Parameter Value
        detail: The request body contains an invalid request parameter value.
        code: 400-08
    invalid-request-parameter-value-with-errors:
      description: Representation of a 400 Bad Request response where the request contains a invalid query or path parameter value.
      value:
        type: https://problems-registry.smartbear.com/invalid-request-parameter-value
        status: 400
        title: Invalid Request Parameter Value
        detail: The request body contains an invalid request parameter value.
        code: 400-08
    unauthorized:
      value:
        type: about:blank
        title: Unauthorized
        detail: Access token not set or invalid. The requested resource could not be returned
        status: 401
    forbidden:
      value:
        type: about:blank
        title: Forbidden
        detail: The resource could not be returned as the requestor is not authorized
        status: 403
    not-found:
      value:
        type: about:blank
        title: Not found
        detail: The requested resource was not found
        status: 404
    unprocessable-entity:
      value:
        type: about:blank
        title: Unprocessable Entity
        detail: The request was malformed or could not be processed.
        status: 422
    server-error:
      value:
        type: about:blank
        title: Server Error
        detail: The server encountered an unexpected error
        status: 500
    insufficient-storage:
      value:
        type: about:blank
        title: Insufficient Storage
        detail: The server is unable to store the representation needed to successfully complete the request
        status: 507